improve windows fault trapping and reporting
Summary:
there are several different things we need to set to have a
stronger chance that our process will terminate when it faults, rather
than popping up a dialog box in some weird service context: set those.

For unhandled SEH exceptions, CaptureStackBackTrace() doesn't seem
to successfully unwind through the SEH trampoline, so we need to
add an auxilliary function that will return the set of frames given
and system exception pointer.

Let's ensure that we initialize the symbol handler before invoking
any of the debug helpers; it seemed to help increase the chances of
getting reliable symbole information.

Increase the size of the possible symbol name; demangled C++ names,
especially those for STL containers are very long; if the buffer is
not long enough, the symbol resolver just sets the name to an empty
string.

Add the error reason in place of the symbol name if we're unable to
resolve it.

Reviewed By: farnz

Differential Revision: D4201457

fbshipit-source-id: 384de72d7f49f74fb7c92bdbfea200e27d87a473