watchman_file::getName -> return string_piece
Summary:
Running down an intermittent problem in the win32 build
and I'm concerned that our embedded w_string trickery is causing problems
by making a false promise to consumers of getName() that the string will
be valid for the lifetime of the reference.

Update: confirmed that this was the problem after beefing up our error
reporting; seemed to only intermittently impact windows, presumably due to
destruction ordering and allocator layout differences.  Stack trace is below.

To avoid this, redefine this method to return a string_piece, forcing
callers to allocate storage or otherwise consider the lifetime.

To facilitate this I've moved around a couple of string helper functions so
that they operate on string_pieces.

I had to update the subscription test: now that we're more reliably tearing
down roots, we can deliver cancellation notices more reliably, and this was
causing an intermittent failure when the subscription payload with a missing
files key was processed in this test.

```
2016-11-17T22:32:54,032: [client=0000020B66990360:stm=0000020B65C1CF70] Unhandled win32 exception.  Fatal error detected at:
```

Reviewed By: farnz

Differential Revision: D4201462

fbshipit-source-id: a35462f8adbebcc19610005c9ad2dfd2205bed43