remove client shutdown hack
Summary:
Don't try to send a signal to a blocked client thread,
or the listener thread.

This was a hacky attempt at working around a blocking problem
during shutdown that we (can) now solve a different way.

This is part of it: knock the listener out of the accept loop
on shutdown.  The remaining part is not requiring that we
enable blocking mode when we write data from the serializer
layer.  I've seen a few stack traces where this signalling
approach clearly wasn't working anyway, so it's not terrible
to remove that portion without proposing a better fix in
the same diff.

Reviewed By: farnz

Differential Revision: D4203947

fbshipit-source-id: f33226f67251494e1d1acfbaf007df627aeaa7da