refactor windows path conversion
Summary:
Use std::wstring in place of explicitly allocated WCHAR pointers.

Note that at least one of the code paths does a slightly wasteful double
allocation so that it can return a strdup'd buffer.

That's "fine" for now; in a follow on diff, I will change the API so that it
can return the w_string directly.

Reviewed By: farnz

Differential Revision: D4205567

fbshipit-source-id: 3f35c82406b6f4b555a0dbd5c1ca72bbf32a3120