add watchman::to<std::string> helper function
Summary:
This reuses the w_string::build() machinery to implement
a function that is similar to `folly::to<std::string>`; the variadic
arguments are appended to a container of the specified type.

While we primarily use w_string for our filename paths, there are a number of
error handling flows that currently use a pretty ugly
`ignore_result(asprintf())` construction to pass a formatted error message back
out to the caller.

I'd like to eliminate this `asprintf` usage and replace it with exception
throwing.  It feels simplest to move a string generated by this function into
such an exception.

With that done, we can also remove the asprintf port that we have floating
around for windows.

Reviewed By: farnz

Differential Revision: D4209509

fbshipit-source-id: f68805d94b614d51044dba23273419adb3e32baf