remove data race from json serializer
Summary:
I was seeing `PERF: (null)` lines getting logged for slow happenings and wanted to figure out why the `json_dumps` call was failing.

I traced this to the serializer code recursion protection.  In this situation
we were serializing the same json object to string buffers from two threads
concurrently. This code isn't and cannot be made thread safe without doing
something like maintaining a set of seen references.  We don't perform
recursion detection in the bser serializer, where we do the bulk of our
serializing, so I don't see the value in this -> removed it.

While I was in there, I made the reference count an atomic variable so that we
can at least have safe inc/dec ops on that.

Reviewed By: farnz

Differential Revision: D4250083

fbshipit-source-id: 1dbfb71e46a26c1f4ba280c63c89f8b0098d623b