avoid making two w_strings in this codepath
Summary:
Since we already know the offset we want, we can just
use a string piece and build the correct final string.

I added a helper for profiling here too; if you set `bench: true`
in your query spec, the query will be run 100 times in a loop
before really running and returning it.  This is helpful to
eliminate some of the noisy overheads during query dispatch.

Reviewed By: sid0

Differential Revision: D4261518

fbshipit-source-id: a653b4fe0d4bf4f535584c94632232f03a2f2b9d