port these php tests to python
Summary:
For unknown reasons these tests caused a TSAN instrumented
build to fail to start up.  I suspect some weird environmental issue
with hhvm (which we use in place of `php`) as a parent of the
spawned watchman processes in these tests.

This slays a dragon that was blocking converting more of the python
tests over; the suspend/resume methods are now in the python
test machinery.

Reviewed By: farnz

Differential Revision: D4254948

fbshipit-source-id: 7205546a6e19e40550f16131845720321700fd77