we don't need strict symlink checks for the state dir
Summary:
simpkins noticed that this broke his symlink-redirected /var dir with the 4.8 pre-release.

This diff adds a flag to our opendir wrapper to skip the strict checks and just
use regular `open` or `opendir` depending on the platform, then makes the state
dir opening code pass false for that flag.

Reviewed By: simpkins

Differential Revision: D4279666

fbshipit-source-id: 88e798a27126961fb0a1c7e144edf9dcd4d378ee