fix regression in pathCat
Summary:
Prior to c++ifying this code, the path cat function would
be smart enough to not emit "/" for empty strings.  This got lost
along the way and was breaking hh_server running against our www
repo; hh_server would pass `relative_root: ""` and because of the
broken path cat function we would interpret that as "/" and cause
all of its queries to never match anything.

Reviewed By: sid0

Differential Revision: D4282990

fbshipit-source-id: 636aaabf950f00481260e1f7ef1a878b93cc2722