expose the client pid in our stream layer
Summary:
While troubleshooting the source of some queries it is valuable to be
able to determine the process id of the peer, so add a method that returns the
pid of the peer if known.

We already had code to query the peer credentials on Linux and Mac, so this
just moves it to the constructor and makes the existing ownership check
a simple accessor for that that, then adds a method to return the process
ID from that same state.

Reviewed By: sid0

Differential Revision: D4282299

fbshipit-source-id: fcfa7a10723d088b4a33a756110933dc39ee0605