standardize on / dir separators on all platforms
Summary:
This removes the WATCHMAN_DIR_SEP constant and instead just forces us
to use / as a directory separator, even on Windows.

This is for our internal representation of paths only; when we translate our
internal multi-byte string to WCHAR we will replace the separators with the
typical windows backslash.

In the future, I want to clean up our suite of functions that operate on
paths and introduce a more strongly typed way of working with them.

For now, this seems sufficient to catch the edges.

Note that this causes watchman to always emit relative file lists with forward
slashes on Windows.  Previously those paths would use backslashes.  I believe
that this is benign because Windows doesn't really care, but may tickle
overzealously pedantic path libraries in some environments.  I don't know if
such a thing exists.

Adding mostafaeweda for Nuclide as the most likely client that might care about this

Reviewed By: farnz

Differential Revision: D4334456

fbshipit-source-id: c05ddf30a6808179cceb5869dccf739feaed8953