fix some bad logging
Summary:
Trying to run down some misbehavior and found that we were logging the parent directory name in this case rather than the actual directory name.

Fix up the logging, and rename this variable to make it clearer that it is the parent and not the dir object associated with the filename being examined.

Reviewed By: sid0

Differential Revision: D4342833

fbshipit-source-id: 119ef9e0d0ddc3f8e7217e123c8defa21c0a3117