speed up directory crawling on windows
Summary:
this takes advantage of the pre-fetched file information that is
available to us from the particular flavor of directory walking that we use on
windows; we've already fetched all of the stat-like information that we
request in our `lstat` implementation (see `watchman/winbuild/stat.cpp`),
so we can fill out that data in our readdir API and set the has-stat flag
to true.

This causes the directory crawler to avoid making an additional `lstat` call.

I haven't benchmarked this on a real system yet, but I suspect the gains are
likely to be larger than you might think; my gut feeling is that by pre-fetching
this information with the buffer size that we use, we are likely a good bit
more expensive than the simplist `FindFilesXXX` API, only to thrash some caches
with the follow up `lstat` calls--those need to open handles and query two separate
sets of information.

So I'm hoping that this has a pretty decent gain.

Reviewed By: farnz

Differential Revision: D4338468

fbshipit-source-id: cffc0b7cebf97d4e07614d05a79ee4c6a187525a