fixup pending list linkage mutation
Summary:
The root cause of the divergence that we were seeing on Linux was
this; in the case where we'd append one list to another, we were not clearing
the prev pointer before inserting it into its new owning pending list.   When
we later went to prune obsoleted nodes, we would end up with garbage linkage
that would cause important nodes to not be visible in list and thus would not
get observed by processPath.

In addition, there was a funky use-after-move-assignment issue that could lead
to a similar issue.

Reviewed By: sid0

Differential Revision: D4348232

fbshipit-source-id: 54916832bd814bb122b50d06f4dfda9182e15e78