allow building runtests.py with buck
Summary:
This is the old test runner -- it was incredibly useful to have around because it allows for neat features like keeping state if things fail, and attaching a debugger to watchman.

We build it as a `python_binary` because we don't yet know how to talk to the standard unittest infrastructure.

Reviewed By: wez

Differential Revision: D4351509

fbshipit-source-id: e6bd658dca95cd9dc453669f317dc06af79f5c18