add w_string_piece::operator<
Summary:
Similar to its counterpart in w_string.  This allows using a string
piece as a key in an ordered_map (you will need to ensure that the string piece
has stable storage for the lifetime of the corresponding value though; you can
do this by embedding the key in the value).

Reviewed By: sid0

Differential Revision: D4352337

fbshipit-source-id: 112a6c45ef3738f1d5943558a94978bc5c5e7672