subscriptions: add a new API to flush subscriptions
Summary:
It is often useful for clients to be able to force a subscription sync, ignoring settle periods and other such things. At the moment there's no way for a client to do that -- this API provides clients an easy way.

The contract is as follows:

* `sync_timeout` is required.
* `subscriptions` is optional -- by default it will sync all subscriptions matching the specified root.
* when the call to `flush-subscriptions` returns, subscriptions are guaranteed to be up to date as of the time the call was made.

We return any pending subscription data as unilateral PDUs. That makes it easy for any code currently processing subscriptions to keep doing what it does.

Reviewed By: wez

Differential Revision: D4351547

fbshipit-source-id: 957caf643943ff32aee1656b53ba261ce9ce726b