Don't swallow exceptions when reading from Watchman socket.
Summary:
During deserialization, the SerializedStreamMessageSupplier was swallowing exceptions
and returning null.  It was doing this to adhere to the com.google.common.base.Supplier
contract.  Instead, use the java.util.concurrent.Callable interface, which is
functionally identical, but has a signature that supports checked exceptions.

Added two unit tests to confirm that errors are propagated to the caller.
(Running those tests with code that swallows exceptions causes the
ListenableFutures returned by the WatchmanConnection to hang.)