Don't swallow exceptions in WatchmanConnection.
Summary:
During deserialization, the SerializedStreamMessageSupplier was swallowing exceptions
and returning null.  It was doing this to adhere to the com.google.common.base.Supplier
contract.  Instead, use the java.util.concurrent.Callable interface, which is (essentially)
functionally identical, but has a signature that supports checked exceptions.

Since the supplier is running in an executor, it is essential that the executor *not* swallow all exceptions, since that makes it immune to canceling by InterruptedExceptions.
Closes https://github.com/facebook/watchman/pull/412

Reviewed By: styurin

Differential Revision: D4387876

Pulled By: mkillianey

fbshipit-source-id: 37bf7baaf6fcbfca60f2497a2d35eb5a2de99f70