For the Java client, allow the user to supply the path to a watchman binary.
Summary:
In the Java client, the WatchmanSocketBuilder finds the first watchman binary
on the user's PATH and uses it by default.  This change allows the user of the
watchman library to pass in a path to an arbitrary executable.  In theory, this
allows a user of this library to use a version of `watchman` that isn't the
first instance on their path, or to use a binary that isn't called `watchman`.

Testing:
Manual testing.