In the Java client, allow the user to select a watchman binary
Summary:
In the Java client, the WatchmanSocketBuilder finds the first watchman binary
on the user's PATH and uses it by default.  This change allows the user of the
watchman library to pass in a path to an arbitrary executable.  In theory, this
allows a user of this library to use a version of `watchman` that isn't the
first instance on their path, or to use a binary that isn't called `watchman`.

Testing:
Manual testing.
Closes https://github.com/facebook/watchman/pull/422

Reviewed By: wez

Differential Revision: D4418602

Pulled By: mkillianey

fbshipit-source-id: 1dc6ba15ba00632424264a6343881eb315522312