refactor spawning into ChildProcess class
Summary:
This refactors a couple of call sites that spawn child processes to
use a new ChildProcess class.

This class is similar to the python subprocess class and thus also the
folly::Subprocess class.  It is a little different in that we're continuing to
target our posix_spawn compatible approach.

The ChildProcess class makes it difficult to leak a child process and helped
uncover a couple of bugs in the trigger command where we could forget and never
reap a child.

The main motivation for this is enabling reading output from the child process,
which is something that we'll need to do when we add more SCM integration.

Reviewed By: sid0

Differential Revision: D4402785

fbshipit-source-id: 9ffb9bc95ba12197ac96b7d0d9f152992e68f2c7