remove socket activation from launchd.plist
Summary:
This will help prevent a situation like
https://github.com/facebook/watchman/issues/358 from having a similar impact in
the future.

Launchd will still launch watchman at startup and keep it running if it falls
over, but should get out of the way when we upgrade in the future.

I'll look at adding some explicit unloading of the plist at installation time
in the homebrew recipe.

Reviewed By: sid0

Differential Revision: D4417672

fbshipit-source-id: b4d4597107f26f2e4e51ecd540dfb36586cfa44c