introduce FileResult abstraction
Summary:
This enables the eden watcher to provide data to the query evaluator
without having to build out a set of watchman_dir and watchman_file objects.

This is a fairly mechanical change to replace uses of watchman_file with
FileResult and provide an implement of FileResult for InMemoryView that just
wraps through to the underlying watchman_file instance.

For the eden watcher, the FileResult implementation uses the FileInformation
that we get back from eden.  It doesn't provide all possible fields yet; those
will be fleshed out as the integration proceeds.

Reviewed By: simpkins

Differential Revision: D4373691

fbshipit-source-id: 95e791013048b35587fc6490da5d9c03b64462be