implement eden path generator
Summary:
This implementation only supports depths of 0 and -1 for the path
generator because we translate to globs and ask eden to execute those.

We could support other depth constraints if we added support for that sort
of query to eden.  We'll punt on that for now and see if anyone asks for
this before adding it.

Reviewed By: simpkins

Differential Revision: D4373994

fbshipit-source-id: ff9862dbc9f74d5645d606cec90e05f2ff9b4cf2