implement eden suffix generator
Summary:
Wherein we discover that we can implement almost every non time based
query in terms of glob.

This translates the suffix list into a recursive glob and asks eden to execute it.
It differs from the InMemoryView implementation in two important ways:

1. Eden matches the glob in a case sensitive fashion whereas Watchman InMemoryView
   defines suffixes as case insensitive.
2. Eden walks the tree on demand so the cost of the query is O(repo), whereas
   Watchman InMemoryView builds an index and the cost of the query is
   O(number-of-files-with-suffix)

I'm fine with both of these differences; we can look at adding an index/cache
of suffix information into eden if we truly need it, and the case sensitive
thing is not a major issue.

Reviewed By: simpkins

Differential Revision: D4374077

fbshipit-source-id: 35f816586ae76cbd411afafbb925df6aed1e91dc