Open source Watchman C++ client library
Summary:
Move the Watchman C++ client library out of HHVM and into the Watchman distribution proper and tidy up for open-sourcing.

The open-source build of Watchman should only include the client library by default only if Folly is available. Configure will hard fail if `--enable-cppclient` is used and folly is not present.

The open-source build of Watchman has now gained a hard dependency on libtool. This could kind of be softened but the dependency would still be present when running `autogen.sh` and this will probably spit out a bunch of whiney error messages with such a setup.

Reviewed By: wez

Differential Revision: D4292974

fbshipit-source-id: 589b5b25bb9487117e7a2b03d86c2e58b74d062f