Fix crash on error getting socket from Watchman
Summary:
To get the default socket name from Watchman we execute the CLI binary and get the result. If something goes wrong an exception is typically thrown, however this can cause us to exit the current code block without waiting for the Subprocess to complete. This causes Subprocess to assert on destruction killing the whole process.

So, we now always wait on the Subprocess which is benign if the process is already finished.

Reviewed By: wez

Differential Revision: D4297859

fbshipit-source-id: e6c142a0c29c376c706dd4c125dfa4202b8523ae