implement since generator for eden
Summary:
Connects the time based generator to the journal API inside eden.

A couple of things to note here:

1. We're delegating the current tick and root number concept to eden;
   these map to the sequenceNumber and mountGeneration fields in
   the eden JournalPosition data type.  This diff connects these
   fields to the appropriate methods of the view.

2. Fresh instance result sets with eden have very slightly different
   semantics from InMemoryView.  The InMemoryView semantics are to
   yield a fresh instance result each time the clock string differs
   in the pid or rootNumber positions.  The way this maps to Eden
   means that we will yield a fresh instance result set whenever
   either watchman is restarted or Eden is remounted.  This is broadly
   speaking OK although it is desirable to only yield a fresh instance
   result set when Eden is remounted, as the watchman process id change
   doesn't require a fresh instance result set with the Eden view.

I've added a very basic ctime and otime clock value to the Eden FileResult
implementation so that we don't crash when calling process_result; this is
sufficient to yield a correct list of names, but not sufficient to reason about
relative ordering of changed files.  Eden doesn't currently provide accurate
information for this purpose, so our hands our tied in this regard for the
moment.

Reviewed By: farnz

Differential Revision: D4374481

fbshipit-source-id: b28a69a0dfe9b60bb4386c90e9defd763fa81da6