add Result type
Summary:
I've been wanting something like this for a little while and it will
help with the light Futures implementation that follows.

Result is most similar to folly::Try but is also slightly influenced by the
rust Result type.

It can hold a value or an error or be empty.

Reviewed By: farnz

Differential Revision: D4469700

fbshipit-source-id: 498b37d29649ee43b2517134094dda24a952d400