add basic Future implementation
Summary:
The intent is to bridge the gap between std::future and folly::Future
without taking on all of the dependencies of folly (we're not in a position to
do this yet, maybe later this year).

Introduces Promise and Future types.  Futures can be chained together using
`.then` to build a staged processing pipeline.  A collection of Futures
can be converted to a future and collected via collectAll.

This implementation doesn't have all of the fancy bells and whistles of its
folly counterpart; the lambda used in the `.then` clause must accept a
`Result<T>&&` or things will likely explode at compile time with an inscrutable
error message.

This is a relatively dumb implementation and won't perform as well as the
aggressively tuned implementation in folly::Future, but it will suffice for our
purposes in watchman, where we want to compose some I/O intensive operations.
The cost of the I/O is expected to dwarf the cost of the Future book keeping.

Reviewed By: farnz

Differential Revision: D4469920

fbshipit-source-id: 3311fabde5eef58264abf4133f039d41df3fdd41