allow query result rendering to use Futures
Summary:
Sets up the plumbing to allow for concurrency/asynchrony in producing
the results for a watchman query.  I've used the readLink FileResult method as
the guinea pig (although it still does not actually do anything async).

There are now two code paths in the result rendering code; one where no futures
are involved and where we can directly emit the results (just as before), and
one that knows how to handle futures.  I've split these code paths to avoid
some of the heap overheads associated with the Future book-keeping.  For the
very large result sets that are common in our usage, this avoids a great many
heap allocations in the common case.

The future aware path looks a little more complex than is strictly needed for a
similar reason; we want to avoid creating a set of futures for each match and
keeping them all around until the whole query completes.

Reviewed By: farnz

Differential Revision: D4470002

fbshipit-source-id: e1ced36c2c932b2810f37765a8dd63c7676ed37c