add LRUCache class
Summary:
The intention is to use this to cache computation of SHA1 content
hashes for the InMemoryView.

The cache uses a hash map to look up an item, and an intrusive doubly linked
list (really a tailq) to maintain the order for eviction purposes.  This
results in most cache operations completing with a couple of O(1) operations
rather than any logarithmic operations.

There's a slightly unsual get() function variant that constrains the number of
concurrent requests for the same key; in this pattern, the client passes both
the key and a function to compute the value for that key if the key is missing.
This allows the cache to perform the computation just once even if there are
multiple requests for the same key.  For example: multiple watchman clients
have subscribed for content hashes and a single change notifies both about
a large number of files.  This allows us to (re)compute the hash for a given
file just once and pass the results to both clients.

Reviewed By: farnz

Differential Revision: D4479070

fbshipit-source-id: 0e2d6af4695aaa4ada8b67388725ea75ede18737