add content hash as an available field
Summary:
This uses the Future and LRUCache plumbing from earlier in the diff
stack to provide a cache for the computed content hash of files in the root.

This is implemented only for InMemoryView at the moment; a follow-on diff will
add the Eden equivalent.

This is purely a demand-fetched implementation at the moment; watchman will
only compute/cache content hashes when a query explicitly requests it.

We cache using the filename, size and mtime as the key to avoid having to think
too hard about invalidations if a given file is changing very rapidly.

A follow-on diff will add some heuristics for cache warming.

I've tried to avoid pulling in openssl on Windows and macOS.  On macOS there is
a SHA1_Init() interface compatible SHA1 library available.  On Windows we use
the crypto API.  On other systems, autoconf looks for libcrypto and adds it to
the linker parameters if present.  The autoconf checks are intentionally a bit
weak to avoid forming a strong dependency on macOS.  Let's see how they do in
the legocastle run.

Reviewed By: farnz

Differential Revision: D4500739

fbshipit-source-id: e5c3ff82cc217cdf55373b02ffc16089e6e897ce