refactor errmsg -> throw QueryXXXError
Summary:
Now that we have exceptions we can tidy this up, so here we are.

This replaces the nasty looking `ignore_result(asprintf(&res->errmsg, "foo %s", "bar")))`
pattern with `throw QueryParseError("foo ", "bar");`

There were two of these `errmsg` fields in the query stuff; one for parse
errors and another for execution errors.  I've introduced a class for both of
these and adjusted the code to call the appropriate flavor.

Reviewed By: simpkins

Differential Revision: D4527192

fbshipit-source-id: 2e24b39c190cd7a4b055e563283a6713da346d38