switch query context from ClockPosition -> ClockSpec
Summary:
a prerequisite for holding the fatter clock values for SCM aware
queries; the idea is that we can store additional contextual information in the
ClockSpec (the since-mergebase and the mergebase-with fields).

Today we assume that we can parse a simple string and return a simple string
value for the clock based on a ClockPosition only.  This diff cuts us over to
using a ClockSpec instead and adds a `toJson` method so that we can transform
the spec into a json representation.

At this time we're still just returning simple string clocks, but in a
subsequent diff we can add more information to these things.

Reviewed By: farnz

Differential Revision: D4628677

fbshipit-source-id: 10fa8811789ec6789670cf0dc1b060ae0765ad70