add basic Mercurial scm awareness to since queries
Summary:
This diff allows adding an `scm` constraint to a query.  This is implemented by
making the clockspec parser understand a "fat" clock, which is really just a
clock with more information inside it than just a string.

A fat clock is represented as a json object rather than just a string.  These
clocks are identical:

```
"c:0:0"
{}
{"clock": "c:0:0"}
```

This also makes it more convenient to specify a "null clock" value; just pass an
empty object.

For SCM awareness we've allowed for the clock to contain an `scm` property with
some parameters to influence the query.  For example, the client can pass in
something like this on their initial query:

```
"since": {
  "scm": {
    "mergebase-with": "master"
  }
}
```

this causes watchman to compute the mergebase with the specific commit.

If the mergebase is the same, then the query continues as a normal `since` query.

If the mergebase has changed, rather than computing a large list of everything
that got changed between the old revision and the new one, watchman will instead
just return the list of files that have been changed since the new mergebase,
filtered by the query expression.

The goal of this feature is to minimize the size of the resultset for clients
that are aware of source control.

In the query response for such a query, watchman will populate the scm information:

```
{
  files: ["foo", "bar"],
  clock: {
    clock: 'c:123:123:123',
    scm: {
      'mergebase': 'HASH',
      'mergebase-with': 'master',
    }
  }
}
```

Note that the the clock value in the result set is a "fat" clock.  Generating the
fat clock value is handled by the `ClockSpec.toJson()` method; it will generate one
if the clockspec contains more than just a ClockPosition.  This can only happen
if the client initiated a query using a fat clock spec as an input, so we should
be good from a backwards compatibility standpoint.

On a subsequent query, the client will provide the `clock` parameter as the value
of the `since` field as they would do normally.  Passing back the fat clock value
causes the next query to continue to use the scm parameters:

```
since: {
  clock: 'c:123:123:123',
  scm: {
    'mergebase': 'HASH',
    'mergebase-with': 'master'
  }
}
```

To facilitate this, I've introduced a small SCM abstraction.  The interface requires
two methods to be implemented; one to return the current mergebase and another to
return the set of !ignored files changed since a specified mergebase.

The QueryableView is responsible for returning an instance of the SCM object.
Putting it at the view layer allows us to very specifically target Eden and
avoid running external commands for the eden watcher.

This diff only implements this functionality for Mercurial.  I plan to extend it
to handle git in a follow-on commit.

Similarly, I plan to add the equivalent functionality for subscriptions in another follow-on commit.

Reviewed By: bolinfest

Differential Revision: D4533545

fbshipit-source-id: 4093c30954da6be6f75a820577f6deed96156fdc