Mercurial: add cache for merge base information
Summary:
This adds a cache of merge base data.  The cache is invalidated by a
change of the mercurial `dirstate` file; this happens whenever an explicit
source control operation takes place on the working copy and triggers whenever
a commit is made or an update operation has taken place (and likely a bit more
frequently than that).   This makes it a quick and easy way to gate executing
mercurial commands to discover the merge base information.

Note that symbolic commitIds passed in as the mergebase-with parameter can
change their target without dirstate being updated.  For the purposes of the
feature that this is used for in watchman, this is "ok".

We invalidate the entire merge base cache whenever the dirstate file changes.

Reviewed By: farnz

Differential Revision: D4634394

fbshipit-source-id: 1bde83e0469948dc440c0ba90cd16bbd8545b6a7