tighten up accessing the view around a recrawl
Summary:
We have a data race where clients can attempt to operate on a view
just as a recrawl is in progress.  This is causing some of our integration
tests to be flaky.

This diff introduces a `view()` method on the root that acquires a read
lock on the recrawl info.  This ensures that we can't read a nullptr
view during the switcheroo that happens in setting up the new root.

It doesn't magically guarantee that the returned root is the active one,
but at least it is a live object and not a nullptr.

Reviewed By: farnz

Differential Revision: D4653767

fbshipit-source-id: 6cfc99e278cbc083cd61b1ed683cbc677f8f300f