retry command invocation for ECONNREFUSED after spawn
Summary:
We've been seeing something that looks like this is a problem on our sandcastle infra.

On that tier we have configured the site spawner that we use on devservers, but
because that tier has no owning user the spawner falls through to
asynchronously starting `watchman --foreground` in the background.  The theory
is that we're racing with its start and hitting ECONNREFUSED because we try to
connect to its socket before it's started to listen on the socket.

To compensate for this, this diff adds a simple retry loop; we'll keep trying to
reconnect for suitable error codes until we've made 10 attempts.   In between
each attempt we'll do a short sleep with exponential backoff.

Reviewed By: farnz

Differential Revision: D4799606

fbshipit-source-id: 7fee81d8b8628745b26247a700177a97cd20977f