adjust since generator behavior w/ unix timestamps
Summary:
Our friends at Microsoft report that the behavior when using
since queries with timestamps is not greedy enough.

Since timestamps are a racy way to reason about changes in the
recency index, we should should prefer to over-notify the client.

For example, it is possible for a file to be changed, a watchman
query to be issued and then a file changed and another query issued
all within the same second.  In this circumstance we should prefer
to over notify the client and tell them about both files.

Prior to this diff we'd only notify about the first file and the
notifcation about the second file would be lost.

The workaround for a sensitive client is to subtract 1 from the
timestamp in this situation.  That's a PITA so we just do that
on the server side.

It is still very much more preferable for the client to use
clock notation instead of timestamps though!

Reviewed By: sid0

Differential Revision: D4988186

fbshipit-source-id: b4515a83aad1123e1f8cc6f8b31aa3bec8b3fe74