explicitly set need_setup=0
Summary:
Well, it's weird.  Sometimes CMD chokes on this not being
defined, and sometimes it doesn't.  It did yesterday, and I needed
this diff.  Today it doesn't.  The difference being that I rebooted
a couple of times since yesterday.

I've also added an arcanist quoting fix in here; I'll split this out once I've
managed to successfully diff this up.

Reviewed By: sid0

Differential Revision: D5032049

fbshipit-source-id: 4a696b17c82b7f0dba1ddd6f883d6f01c64b8dca