fixup some gcc 7.1 issues
Summary:
We've had reports that gcc has a new warning about case
statements that fall through.  In addition, something has changed
in the implicit dependencies that require us to explicitly include
`<functional>`.

There are some attribute-y things we could do to annotate this
more deliberately, but I'm punting on that for the moment.

Refs: https://github.com/facebook/watchman/issues/467

Reviewed By: farnz

Differential Revision: D5087816

fbshipit-source-id: 45392a769c063966291ad4ab6b4347bbb2867802